requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.0.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
flake8==6.1.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
Pillow==10.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
